+++ front-end/package.json
@@ -26,7 +26,6 @@
     "expo-status-bar": "~1.12.1",
     "expo-system-ui": "~3.0.7",
     "expo-web-browser": "~13.0.3",
-    "json-server": "^1.0.0-beta.2",
     "nativewind": "4.0.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
