+++ front-end/src/app/(tabs)/index.tsx
@@ -1,11 +1,10 @@
 import { Text, View, ScrollView } from "react-native";
 import { Header } from "../../components/header";
-import { Banner } from "../../components/banner";
-import { LastCourt } from "@/src/components/lastCourt";
+import Banner from "../../components/banner";
+import LastCourt from "@/src/components/lastCourt";
 
 import Constants from 'expo-constants'
 import { FilterSport } from "../../components/filterSport";
-
 
 const statusBarHeight = Constants.statusBarHeight;
 
@@ -22,9 +21,7 @@ export default function Inicio() {
         <Text className="text-xl">Vamos Jogar hoje?</Text>
         <FilterSport />
         <Banner />
-        <Text className="text-xl font-semibold mt-3">Ultimas Quadras Visitadas</Text>
-        <LastCourt/>
-
+        <LastCourt />
       </View>
     </ScrollView>
   );
